Renames comicmodels to challenges
