+++ app/challenges/emails.py
@@ -2,7 +2,7 @@ from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.core.mail import send_mail
 
-from comicmodels.models import ComicSite
+from challenges.models import ComicSite
 
 
 def send_challenge_created_email(challenge: ComicSite):
