+++ app/challenges/templates/challenges/comicsite_list.html
