+++ app/challenges/templates/challenges/comicsite_update.html
