--- app/comicmodels/__init__.py
@@ -1,1 +1,0 @@
-default_app_config = 'comicmodels.apps.ComicModelsConfig'
