+++ app/evaluation/models.py
@@ -5,7 +5,7 @@ from django.core.exceptions import ValidationError
 from django.db import models
 from social_django.fields import JSONField
 
-from comicmodels.models import ComicSite
+from challenges.models import ComicSite
 from comicsite.core.urlresolvers import reverse
 from evaluation.emails import send_failed_job_email
 from evaluation.validators import MimeTypeValidator, ExtensionValidator
