+++ app/participants/models.py
@@ -1,7 +1,7 @@
 from django.conf import settings
 from django.db import models
 
-from comicmodels.models import ComicSite
+from challenges.models import ComicSite
 
 
 class RegistrationRequest(models.Model):
