+++ app/resetmigrations.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-for app in uploads teams participants pages evaluation comicmodels profiles jqfileupload
+for app in uploads teams participants pages evaluation challenges profiles jqfileupload
 do
     python manage.py migrate --fake $app zero
     find . -path "./$app/migrations/*.py" -not -name "__init__.py" -delete
