+++ app/tests/comicsite_tests/integration_tests.py
@@ -16,7 +16,7 @@ from django.test.utils import override_settings
 from six import StringIO
 from userena.models import UserenaSignup
 
-from comicmodels.models import ComicSite
+from challenges.models import ComicSite
 from dataproviders.utils.HtmlLinkReplacer import HtmlLinkReplacer
 from pages.models import Page
 from tests.factories import PageFactory, RegistrationRequestFactory
