+++ app/uploads/views.py
@@ -22,8 +22,8 @@ from django.utils.html import escape
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import ListView, CreateView, TemplateView
 
-from comicmodels.models import ComicSite
-from comicmodels.permissions import can_access
+from challenges.models import ComicSite
+from challenges.permissions import can_access
 from comicsite.permissions.mixins import UserIsChallengeAdminMixin
 from comicsite.views import getSite, site_get_standard_vars, permissionMessage
 from pages.models import Page
